+++ docs/history.rst
@@ -2,13 +2,17 @@
 History
 =======
 
+0.1.4 (????-??-??)
+------------------
+* Now an optimizer can either be passed as a string or the TensorFlow op
+* Fix replay memory bug where size would only reach max size - 1
+
+
 0.1.2/3 (2016-17-15)
 ------------------
-
 * Improving meta data and fixing __init__ scripts to load subpackages properly
 
 
 0.1.0 (2016-16-15)
 ------------------
-
 * First release on PyPI.